d/rules: fix dangling symlink of vmware-user
Back in 2:9.4.0-1280544-6 vmware-user* was moved to the
open-vm-tools-desktop package and some follow on fixes moved bits that
were forgotten like the man page.

There still is a symlink in /usr/bin/vmware-user that is forgotten in
the base package and broken unless open-vm-tools-desktop is installed.
Change d/rules to move the symlink as well.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>