Start vgauth before vmtoolsd
VGAuthService needs to be ready when vmtoolsd runs. Certain cases - e.g.
Site Recovery Manager failover - will need vgauth to be up.
Therefore add an After=vgauth.service dependency to open-vm-tools.service

To have vgauth be able start early - and not pull cloud-init back late - it
is also required to drop default dependencies which according to VMware is
fine to do so.

Fixes: https://bugs.launchpad.net/ubuntu/+source/open-vm-tools/+bug/1804287

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>